Moving stat comparison into shared module.
