+++ shared/ui/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/ui/stats/CoreStatsComparisonCard.kt
@@ -1,6 +1,5 @@
-package com.adammcneilly.pocketleague.android.designsystem.stats
+package com.adammcneilly.pocketleague.shared.ui.stats
 
-import android.content.res.Configuration
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -9,10 +8,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.adammcneilly.pocketleague.core.displaymodels.CoreStatsDisplayModel
-import com.adammcneilly.pocketleague.shared.design.system.theme.PocketLeagueTheme
 
 /**
  * Render a comparison of [blueTeamStats] and [orangeTeamStats] for the
@@ -61,37 +58,5 @@ fun CoreStatsComparisonCard(
                 orangeTeamValue = orangeTeamStats.saves,
             )
         }
-    }
-}
-
-@Preview(
-    name = "Night Mode",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-)
-@Preview(
-    name = "Day Mode",
-    uiMode = Configuration.UI_MODE_NIGHT_NO,
-)
-@Composable
-private fun CoreStatsComparisonCardPreview() {
-    PocketLeagueTheme {
-        CoreStatsComparisonCard(
-            blueTeamStats = CoreStatsDisplayModel(
-                score = 1000,
-                goals = 7,
-                assists = 4,
-                saves = 3,
-                shots = 1,
-            ),
-            orangeTeamStats = CoreStatsDisplayModel(
-                score = 8729374,
-                goals = 3,
-                assists = 8,
-                saves = 0,
-                shots = 5,
-            ),
-            modifier = Modifier
-                .padding(16.dp),
-        )
     }
 }
