+++ shared/ui/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/ui/utils/WhenInView.kt
@@ -1,6 +1,5 @@
-package com.adammcneilly.pocketleague.android.designsystem.utils
+package com.adammcneilly.pocketleague.shared.ui.utils
 
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -10,12 +9,12 @@ import androidx.compose.ui.composed
 import androidx.compose.ui.layout.findRootCoordinates
 import androidx.compose.ui.layout.onGloballyPositioned
 
+
 /**
  * This is a custom [Modifier] that observes whether or not a
  * composable has been rendered on the screen, and once it has
  * it will execute the given [callback].
  */
-@Composable
 fun Modifier.whenInView(
     callback: () -> Unit,
 ): Modifier = composed {
@@ -35,3 +34,4 @@ fun Modifier.whenInView(
         }
     }
 }
+
