+++ shared/ui/src/test/snapshots/images/com.adammcneilly.pocketleague.shared.ui.stats_StatComparisonPaparazziTest_renderCompleteStatComparison[useDarkTheme=false].png
