+++ shared/ui/src/test/snapshots/images/com.adammcneilly.pocketleague.shared.ui.stats_StatComparisonPaparazziTest_renderHalfwayStatComparison[useDarkTheme=true].png
